+++ metrics/benches/macros.rs
@@ -21,7 +21,7 @@ impl Recorder for TestRecorder {
     }
     fn increment_counter(&self, _key: Key, _value: u64) {}
     fn update_gauge(&self, _key: Key, _value: GaugeValue) {}
-    fn record_histogram(&self, _key: Key, _value: u64) {}
+    fn record_histogram(&self, _key: Key, _value: f64) {}
 }
 
 fn reset_recorder() {
